\name{tweeDEseq-internal}
\alias{tweeDEseq-internal}



\alias{loglikPoissonTweedie}
\alias{loglikPoissonTweedie2}
\alias{momentEstimates}

\title{Internal 'tweeDEseq' functions}
\description{Internal tweeDEseq functions}

\usage{
loglikPoissonTweedie(p, x, mu, verbose = FALSE, tol = 1e-15)
loglikPoissonTweedie2(p, a, x, mu, verbose = FALSE, tol = 1e-15)
momentEstimates(x)
}

\details{These are not to be called by the user}
\keyword{internal}
