\name{tweeDEseq-internal}
\alias{tweeDEseq-internal}

\alias{loglikPoissonTweedie}
\alias{loglikPoissonTweedie2}
\alias{momentEstimates}
\alias{shapeTrend}

\title{Internal 'tweeDEseq' functions}
\description{Internal tweeDEseq functions}

\usage{
loglikPoissonTweedie(p, x, mu, verbose = FALSE, tol = 1e-15, probs =
FALSE, w = NULL)
loglikPoissonTweedie2(p, a, x, mu, verbose = FALSE, tol = 1e-15, probs =
FALSE, w = NULL)
momentEstimates(x)
shapeTrend(x)
}

\details{These are not to be called by the user}
\keyword{internal}
